backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
pydantic-settings>=2.6,<3.0
email-validator>=2.2,<3.0
supabase>=2.9,<3.0
passlib>=1.7,<2.0
pandas>=2.2,<3.0
plotly>=5.24,<6.0
requests>=2.32,<3.0
python-multipart>=0.0.20,<1.0
